Made changes per our design discussion. ATM requests now raise exceptions and do some field validation. The ATM and ATMRequest classes were seperated into different files. A constants file and an exceptions file were added
