+++ src/utils/constants.py
@@ -1,0 +1,6 @@
+#Urls
+baseUrl = "http://api.reimaginebanking.com"
+
+#Exception Messages
+missingFields = 1
+invalidFields = 2
